refactor(cmd): extract usage printing to usage.go
